// src/generated-api/src/models/DomainFavoritePublisherListItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DomainAnalyticsPublisher } from './DomainAnalyticsPublisher';
import {
    DomainAnalyticsPublisherFromJSON,
    DomainAnalyticsPublisherFromJSONTyped,
    DomainAnalyticsPublisherToJSON,
    DomainAnalyticsPublisherToJSONTyped,
} from './DomainAnalyticsPublisher';

/**
 * 
 * @export
 * @interface DomainFavoritePublisherListItem
 */
export interface DomainFavoritePublisherListItem {
    /**
     * 
     * @type {string}
     * @memberof DomainFavoritePublisherListItem
     */
    addedAt?: string;
    /**
     * 
     * @type {number}
     * @memberof DomainFavoritePublisherListItem
     */
    itemId?: number;
    /**
     * 
     * @type {number}
     * @memberof DomainFavoritePublisherListItem
     */
    listId?: number;
    /**
     * 
     * @type {string}
     * @memberof DomainFavoritePublisherListItem
     */
    notes?: string;
    /**
     * Optional: Include publisher details when fetching with details
     * @type {DomainAnalyticsPublisher}
     * @memberof DomainFavoritePublisherListItem
     */
    publisher?: DomainAnalyticsPublisher;
    /**
     * 
     * @type {string}
     * @memberof DomainFavoritePublisherListItem
     */
    publisherDomain?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainFavoritePublisherListItem
     */
    status?: string;
}

/**
 * Check if a given object implements the DomainFavoritePublisherListItem interface.
 */
export function instanceOfDomainFavoritePublisherListItem(value: object): value is DomainFavoritePublisherListItem {
    return true;
}

export function DomainFavoritePublisherListItemFromJSON(json: any): DomainFavoritePublisherListItem {
    return DomainFavoritePublisherListItemFromJSONTyped(json, false);
}

export function DomainFavoritePublisherListItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): DomainFavoritePublisherListItem {
    if (json == null) {
        return json;
    }
    return {
        
        'addedAt': json['added_at'] == null ? undefined : json['added_at'],
        'itemId': json['item_id'] == null ? undefined : json['item_id'],
        'listId': json['list_id'] == null ? undefined : json['list_id'],
        'notes': json['notes'] == null ? undefined : json['notes'],
        'publisher': json['publisher'] == null ? undefined : DomainAnalyticsPublisherFromJSON(json['publisher']),
        'publisherDomain': json['publisher_domain'] == null ? undefined : json['publisher_domain'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function DomainFavoritePublisherListItemToJSON(json: any): DomainFavoritePublisherListItem {
    return DomainFavoritePublisherListItemToJSONTyped(json, false);
}

export function DomainFavoritePublisherListItemToJSONTyped(value?: DomainFavoritePublisherListItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'added_at': value['addedAt'],
        'item_id': value['itemId'],
        'list_id': value['listId'],
        'notes': value['notes'],
        'publisher': DomainAnalyticsPublisherToJSON(value['publisher']),
        'publisher_domain': value['publisherDomain'],
        'status': value['status'],
    };
}